pyproject.toml
[tool.poetry]
name = "ringmybell"
version = "0.1.0"
description = "A twitter streaming application that uses a relay to ring a bell when the streamed address receives a tweet."
authors = ["Andy McNeil"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
twython = "^3.8.2"
"rpi.gpio" = "^0.7.0"
picamera = "^1.13"
adafruit-circuitpython-ht16k33 = "^4.1.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
ringmybell = "ringmybell.main:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

